+++ spec/unit/attributes_spec.rb
@@ -8,7 +8,7 @@ describe AttributesDSL::Attributes do
   describe ".new" do
     subject { attributes }
 
-    it { is_expected.to be_frozen }
+    it { is_expected.to be_immutable }
   end # describe .new
 
   describe "#attributes" do